fix: re-use formatBytes function
