+++ src/routes/+page.svelte
@@ -6,7 +6,7 @@
   import ProcessTable from "$lib/components/ProcessTable.svelte";
   import ProcessDetailsModal from "$lib/components/ProcessDetailsModal.svelte";
   import KillProcessModal from "$lib/components/KillProcessModal.svelte";
-  import { formatNetwork, formatStatus } from "$lib/utils";
+  import { formatBytes, formatStatus } from "$lib/utils";
   import { themeStore } from "$lib/stores";
   import type { Process, SystemStats, Column } from "$lib/types";
 
@@ -54,13 +54,13 @@
       id: "network_rx",
       label: "Network RX",
       visible: true,
-      format: (v) => formatNetwork(v),
+      format: (v) => formatBytes(v),
     },
     {
       id: "network_tx",
       label: "Network TX",
       visible: true,
-      format: (v) => formatNetwork(v),
+      format: (v) => formatBytes(v),
     },
     { id: "command", label: "Command", visible: false },
     { id: "ppid", label: "Parent PID", visible: false },
